configured ts react app
